backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
Pillow
drf-yasg
reportlab>=4.0.0
